Unify full-featured content editing
