+++ packages/web/src/components/add-entry/NoteForm.svelte
@@ -1,16 +1,9 @@
 <script lang="ts">
-  import type { Component } from 'svelte';
   import type { InboxItem } from '@inbox-rs/rs-module';
-  import { autofocus, autofocusIf } from '../../lib/actions';
-  import {
-    type BuildItemFn,
-    loadMarkdownEditorComponent,
-    type MarkdownEditorProps,
-    shouldLoadMarkdownEditor,
-  } from '../../lib/add-entry-modal';
+  import { autofocusIf } from '../../lib/actions';
+  import type { BuildItemFn } from '../../lib/add-entry-modal';
   import { buildNoteItem } from '../../lib/build-item';
-  import { createCodeKeydownHandler } from '../../lib/code-indent';
-  import { renderMarkdown } from '../../lib/markdown';
+  import MarkdownContentField from './MarkdownContentField.svelte';
 
   let {
     editItem,
@@ -40,15 +33,6 @@
   );
   let description = $state(editItem?.description ?? '');
 
-  let editorMode = $state<'visual' | 'write' | 'preview'>('visual');
-  let MarkdownEditorComponent = $state<Component<MarkdownEditorProps> | null>(
-    null,
-  );
-  let markdownEditorLoadError = $state('');
-  let previewHtml = $state('');
-
-  const handleCodeKeydown = createCodeKeydownHandler();
-
   // A note is captureable as soon as the user has typed a title or any body.
   $effect(() => {
     canSubmit = !!(title || body);
@@ -60,72 +44,6 @@
       { title, body, description },
     );
 
-  // Lazy-load the visual editor on demand: it pulls in TipTap and friends
-  // (~hundreds of KB) and only ~60% of users open the note modal in any
-  // given session. Falls back to the markdown textarea if loading fails so
-  // the user can still capture content.
-  $effect(() => {
-    if (
-      !shouldLoadMarkdownEditor(
-        'note',
-        editorMode,
-        !!MarkdownEditorComponent,
-        !!markdownEditorLoadError,
-      )
-    ) {
-      return;
-    }
-
-    let cancelled = false;
-
-    loadMarkdownEditorComponent()
-      .then((component) => {
-        if (!cancelled) {
-          MarkdownEditorComponent = component;
-        }
-      })
-      .catch((loadError) => {
-        console.error('Failed to load markdown editor:', loadError);
-        if (!cancelled) {
-          markdownEditorLoadError =
-            'Visual editor unavailable. Markdown mode is still available.';
-          editorMode = 'write';
-        }
-      });
-
-    return () => {
-      cancelled = true;
-    };
-  });
-
-  // Render markdown preview asynchronously. Re-checks `body === currentBody`
-  // before applying the rendered HTML so a stale render from a previous body
-  // value doesn't overwrite a fresher one.
-  $effect(() => {
-    if (editorMode !== 'preview') {
-      previewHtml = '';
-      return;
-    }
-
-    const currentBody = body;
-    let cancelled = false;
-
-    renderMarkdown(currentBody)
-      .then((html) => {
-        if (!cancelled && body === currentBody && editorMode === 'preview') {
-          previewHtml = html;
-        }
-      })
-      .catch(() => {
-        if (!cancelled) {
-          previewHtml = '';
-        }
-      });
-
-    return () => {
-      cancelled = true;
-    };
-  });
 </script>
 
 <label class="field">
@@ -138,68 +56,4 @@
   -->
   <input use:autofocusIf={!prefillTitle} type="text" bind:value={title} placeholder="Note title" />
 </label>
-<div class="field note-editor-field">
-  <div class="field-header">
-    <span>Content</span>
-    <div class="editor-tabs">
-      <button
-        type="button"
-        class="tab"
-        class:active={editorMode === 'visual'}
-        onclick={() => (editorMode = 'visual')}
-        disabled={!!markdownEditorLoadError}>Visual</button
-      >
-      <button
-        type="button"
-        class="tab"
-        class:active={editorMode === 'write'}
-        onclick={() => (editorMode = 'write')}>Markdown</button
-      >
-      <button
-        type="button"
-        class="tab"
-        class:active={editorMode === 'preview'}
-        onclick={() => (editorMode = 'preview')}>Preview</button
-      >
-    </div>
-  </div>
-  {#if editorMode === 'visual'}
-    {#if MarkdownEditorComponent}
-      <MarkdownEditorComponent bind:value={body} placeholder="Write your note..." focusOnMount={!!prefillTitle} />
-    {:else if markdownEditorLoadError}
-      <textarea
-        use:autofocus
-        class="code-input"
-        bind:value={body}
-        rows="10"
-        placeholder="Write your note in Markdown..."
-        onkeydown={handleCodeKeydown}
-      ></textarea>
-    {:else}
-      <div class="editor-loading" aria-live="polite">Loading visual editor…</div>
-    {/if}
-  {:else if editorMode === 'write'}
-    <textarea
-      use:autofocus
-      class="code-input"
-      bind:value={body}
-      rows="10"
-      placeholder="Write your note in Markdown..."
-      onkeydown={handleCodeKeydown}
-    ></textarea>
-  {:else}
-    <div class="preview-wrap markdown-body">
-      {#if previewHtml}
-        <!-- eslint-disable-next-line svelte/no-at-html-tags -->
-        {@html previewHtml}
-      {:else if body.trim()}
-        <span class="preview-empty">Preview unavailable</span>
-      {:else}
-        <span class="preview-empty">Nothing to preview</span>
-      {/if}
-    </div>
-  {/if}
-</div>
-{#if markdownEditorLoadError}
-  <p class="info-note">{markdownEditorLoadError}</p>
-{/if}
+<MarkdownContentField bind:value={body} focusOnMount={!!prefillTitle} />
